requirements/base.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-binance==1.0.19
python-decouple==3.8
pytz==2023.3
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework-simplejwt==5.3.0
django-jazzmin==2.6.0
django-model-utils==4.3.1
drf-yasg==1.21.7
drf-standardized-errors==0.12.6